portato float a double per maggiore precisione

+++ fractol.h
@@ -6,7 +6,7 @@
 /*   By: valerio <valerio@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/17 22:48:26 by valerio           #+#    #+#             */
-/*   Updated: 2024/10/18 17:19:54 by valerio          ###   ########.fr       */
+/*   Updated: 2024/10/19 12:08:52 by valerio          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,20 +37,20 @@ typedef struct s_data
 	void	*mlx;
 	void	*win;
 	t_img	img;
-	float	xc; //per zoom (da rivedere)
-	float	yc; //per zoom (da rivedere)
-	float	semiax; // per zoom (da rivedere)
-	float	k;
-	float	**palette;
+	double	xc; //per zoom (da rivedere)
+	double	yc; //per zoom (da rivedere)
+	double	semiax; // per zoom (da rivedere)
+	double	k;
+	double	**palette;
 }	t_data;
 
 typedef	unsigned char	byte;
 
 int		encode_rgb(byte red, byte green, byte blue);
 void	my_pixel_put(t_img *img, int x, int y, int color);
-float	*linspace(float start, float end, int num);
+double	*linspace(double start, double end, int num);
 int 	map(int value, int s0, int e0, int s1, int e1);
-float	**get_palette(float colors[][3], int num_colors);
+double	**get_palette(double colors[][3], int num_colors);
 void	anti_aliasing(t_img *img); //da vedere
 
 int		key_handler(int keysym, t_data *data);
