+++ fractol_window.c
@@ -6,7 +6,7 @@
 /*   By: valerio <valerio@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/17 22:48:37 by valerio           #+#    #+#             */
-/*   Updated: 2024/10/18 16:50:56 by valerio          ###   ########.fr       */
+/*   Updated: 2024/10/19 12:08:07 by valerio          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,7 @@ void fractol_init_window()
 	data.xc = 0;
 	data.yc = 0;
 
-	float	colors[][3] = {{25, 5, 55}, {140, 110, 85}, {60, 160, 130}, {55, 85, 230}, {230, 190, 55}, {255, 0, 0}}; // scrivere una funzione colors_init() che li alloca dinamicamente
+	double	colors[][3] = {{25, 5, 55}, {140, 110, 85}, {60, 160, 130}, {55, 85, 230}, {230, 190, 55}, {255, 0, 0}}; // scrivere una funzione colors_init() che li alloca dinamicamente
 	data.palette = get_palette(colors, 6);
 
 	draw_mandelbrot(&data, ITER);
